Create GB
An optimization model to solve the routing and scheduling problem for offshore wind farms in the paper: Optimisation of maintenance routing and scheduling for offshore wind farms. 
Please Notice: this model is a mixed-integer linear programming problem, which is NP-hard. Without a good initial solution providing a good lower bound, this problem cannot solve within hours.